facturaGenerator/requirements.txt
Flask==3.0.0
reportlab==4.0.7
Werkzeug==3.0.1
num2words==0.5.13
gunicorn==21.2.0
boto3==1.34.28
python-dotenv==1.0.0
